rename : rename directories and moved files
